requirements.txt
discord.py
discord.py[voice]
mysql-connector-python
Halo
configparser
requests
beautifulsoup4
lxml
coloredlogs
ffmpeg
